Update README files and fix system proxy function
- Updated Crates.io badge links in both English and Chinese README files to reflect the correct package name.
- Fixed a type casting issue in the `get_system_proxy` function within `system.rs` to ensure proper return type handling.

+++ src/utils/system.rs
@@ -62,7 +62,7 @@ pub fn get_system_proxy() -> String {
                 0,
                 KEY_ALL_ACCESS,
                 &mut hkey,
-            );
+            ) as i32;
 
             if ret != ERROR_SUCCESS {
                 return String::new();
